Add files via upload
I tried to edit the table like proposed on Tuesday avoiding the doubled information of the unmarked concept in a separate column. I asigned a unique concept number and a concept pair number to each concept. The last column "Hilfsspalte" was only edited to be able to quickly formate and sort the table. Does that make sense?